+++ src/v8/src/libplatform/tracing/trace-writer.cc
@@ -113,7 +113,7 @@ void JSONTraceWriter::AppendArgValue(uint8_t type,
 }
 
 void JSONTraceWriter::AppendArgValue(ConvertableToTraceFormat* value) {
-  int size = value->AppendAsTraceFormat(nullptr, 0);
+  size_t size = value->AppendAsTraceFormat(nullptr, 0);
   if (size) {
     std::unique_ptr<char[]> arg_stringified(new char[size]);
     value->AppendAsTraceFormat(arg_stringified.get(), size);
